+++ src/index.html
@@ -20,7 +20,7 @@
   </footer>
 
   <script type="module" src="main.js"></script>
-
+  
 </body>
 
 </html>
