+++ src/lib/firebase.js
@@ -1,2 +1,2 @@
 export { initializeApp } from 'https://www.gstatic.com/firebasejs/9.9.4/firebase-app.js';
-export { createUserWithEmailAndPassword } from 'https://www.gstatic.com/firebasejs/9.10.0/firebase-auth.js';
+export { getAuth, createUserWithEmailAndPassword } from 'https://www.gstatic.com/firebasejs/9.10.0/firebase-auth.js';
