+++ src/main.js
@@ -1,5 +1,5 @@
-//import './configuration.js'
-import {addUser} from './lib/index.js';
+//import'./configuration.js';
+
 
 import load from './pages/load/load.js'
 import entrar from './pages/entrar/entrar.js'
@@ -10,10 +10,9 @@ const main = document.querySelector('#root');
 const googleBtn = document.querySelector('#google-login');
 const okLoginBtn = document.querySelector('#ok-login-btn');
 const signInName = document.querySelector('#name');
-const signInEmail = document.querySelector('#email').value;
-const signInPassword = document.querySelector('#password').value;
+
 //const signInConfirmPassword = document.querySelector('#confirm-password')
-const okSigninBtn = document.querySelector('#ok-form-btn');
+
 const checkBox = document.querySelector('#checkbox');
 
 const navegar = () => {
@@ -40,8 +39,4 @@ window.addEventListener('load', () => {
     navegar();
 })
 
-okSigninBtn.addEventListener('click', () => {
-    const cadastro = addUser(signInEmail, signInPassword)
-    return console.log(cadastro)
 
-})
